# app/api/user/profile/route.ts
import { NextResponse } from "next/server"
import { auth } from "@/lib/auth"
import { db } from "@/db"
import { users } from "@/db/schema"
import { eq } from "drizzle-orm"
import { z } from "zod"
import bcrypt from "bcryptjs"

const profileSchema = z.object({
  name: z.string().min(2, "Name must be at least 2 characters"),
  password: z
    .string()
    .min(8, "Password must be at least 8 characters")
    .optional()
    .or(z.literal("")),
})

export async function PUT(req: Request) {
  try {
    const session = await auth()

    if (!session || !session.user || !session.user.email) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 })
    }

    const body = await req.json()
    const result = profileSchema.safeParse(body)

    if (!result.success) {
      return NextResponse.json(
        { error: "Invalid input", details: result.error.errors },
        { status: 400 }
      )
    }

    const { name, password } = result.data
    const updateData: { name: string; password?: string } = { name }

    if (password && password.length > 0) {
      // Fetch current user to check password
      const currentUser = await db.query.users.findFirst({
        where: eq(users.email, session.user.email),
        columns: { password: true }
      })

      if (currentUser && currentUser.password) {
        const isSamePassword = await bcrypt.compare(password, currentUser.password)
        if (isSamePassword) {
          return NextResponse.json(
            { error: "New password cannot be the same as the old password." },
            { status: 400 }
          )
        }
      }

      const hashedPassword = await bcrypt.hash(password, 10)
      updateData.password = hashedPassword
    }

    await db
      .update(users)
      .set(updateData)
      .where(eq(users.email, session.user.email))

    return NextResponse.json(
      { message: "Profile updated successfully" },
      { status: 200 }
    )
  } catch (error) {
    console.error("Profile update error:", error)
    return NextResponse.json(
      { error: "Internal server error" },
      { status: 500 }
    )
  }
}